requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
pwinput==1.0.3
Pygments==2.18.0
python-dotenv==1.0.1
rich==13.9.4
